+++ R/pkg-vctrs.R
@@ -56,24 +56,28 @@ vec_cast.wk_wkb.wk_wkt <- function(x, to, ...) {
 #' @method vec_cast.wk_wkb wk_xy
 #' @export
 vec_cast.wk_wkb.wk_xy <- function(x, to, ...) {
+  wk_is_geodesic(x) <- wk_is_geodesic(to)
   wk_translate(x, to)
 }
 
 #' @method vec_cast.wk_wkb wk_xyz
 #' @export
 vec_cast.wk_wkb.wk_xyz <- function(x, to, ...) {
+  wk_is_geodesic(x) <- wk_is_geodesic(to)
   wk_translate(x, to)
 }
 
 #' @method vec_cast.wk_wkb wk_xym
 #' @export
 vec_cast.wk_wkb.wk_xym <- function(x, to, ...) {
+  wk_is_geodesic(x) <- wk_is_geodesic(to)
   wk_translate(x, to)
 }
 
 #' @method vec_cast.wk_wkb wk_xyzm
 #' @export
 vec_cast.wk_wkb.wk_xyzm <- function(x, to, ...) {
+  wk_is_geodesic(x) <- wk_is_geodesic(to)
   wk_translate(x, to)
 }
 
@@ -195,24 +199,28 @@ vec_cast.wk_wkt.wk_wkb <- function(x, to, ...) {
 #' @method vec_cast.wk_wkt wk_xy
 #' @export
 vec_cast.wk_wkt.wk_xy <- function(x, to, ...) {
+  wk_is_geodesic(x) <- wk_is_geodesic(to)
   wk_translate(x, to)
 }
 
 #' @method vec_cast.wk_wkt wk_xyz
 #' @export
 vec_cast.wk_wkt.wk_xyz <- function(x, to, ...) {
+  wk_is_geodesic(x) <- wk_is_geodesic(to)
   wk_translate(x, to)
 }
 
 #' @method vec_cast.wk_wkt wk_xym
 #' @export
 vec_cast.wk_wkt.wk_xym <- function(x, to, ...) {
+  wk_is_geodesic(x) <- wk_is_geodesic(to)
   wk_translate(x, to)
 }
 
 #' @method vec_cast.wk_wkt wk_xyzm
 #' @export
 vec_cast.wk_wkt.wk_xyzm <- function(x, to, ...) {
+  wk_is_geodesic(x) <- wk_is_geodesic(to)
   wk_translate(x, to)
 }
 
@@ -832,6 +840,14 @@ vec_ptype2.wk_rct <- function(x, y, ...) {
   UseMethod("vec_ptype2.wk_rct", y) # nocov
 }
 
+#' @method vec_cast.wk_rct wk_rct
+#' @export
+vec_cast.wk_rct.wk_rct <- function(x, to, ...) {
+  wk_crs_output(x, to)
+  wk_is_geodesic_output(x, to)
+  x
+}
+
 #' @method vec_ptype2.wk_rct wk_rct
 #' @export
 vec_ptype2.wk_rct.wk_rct <- function(x, y, ..., x_arg = "x", y_arg = "y") {
@@ -903,6 +919,14 @@ vec_cast.wk_crc <- function(x, to, ...) {
 #' @export
 vec_cast.wk_crc.default <- function(x, to, ...) {
   vctrs::vec_default_cast(x, to) # nocov
+}
+
+#' @method vec_cast.wk_crc wk_crc
+#' @export
+vec_cast.wk_crc.wk_crc <- function(x, to, ...) {
+  wk_crs_output(x, to)
+  wk_is_geodesic_output(x, to)
+  x
 }
 
 #' @rdname vctrs-methods